+++ src/main/java/stg/onyou/repository/UserClubRepository.java
@@ -5,6 +5,6 @@ import stg.onyou.model.entity.UserClub;
 
 import java.util.List;
 
-public interface UserClubRepository extends JpaRepository<UserClub, Integer> {
-    List<UserClub> findAllByClubId(Integer id);
+public interface UserClubRepository extends JpaRepository<UserClub, Long> {
+    List<UserClub> findAllByClubId(Long id);
 }
